test: move tests to persistence
