+++ persistence/sql/persister_client.go
@@ -11,18 +11,13 @@ import (
 	"github.com/ory/x/sqlcon"
 )
 
-func (p *Persister) GetClient(ctx context.Context, id string) (fosite.Client, error) {
-	return p.GetConcreteClient(ctx, id)
+func (p *Persister) GetConcreteClient(ctx context.Context, id string) (*client.Client, error) {
+	var cl client.Client
+	return &cl, sqlcon.HandleError(p.Connection(ctx).Find(&cl, id))
 }
 
-func (p *Persister) CreateClient(ctx context.Context, c *client.Client) error {
-	h, err := p.r.ClientHasher().Hash(ctx, []byte(c.Secret))
-	if err != nil {
-		return err
-	}
-	c.Secret = string(h)
-
-	return sqlcon.HandleError(p.Connection(ctx).Create(c, "pk"))
+func (p *Persister) GetClient(ctx context.Context, id string) (fosite.Client, error) {
+	return p.GetConcreteClient(ctx, id)
 }
 
 func (p *Persister) UpdateClient(ctx context.Context, cl *client.Client) error {
@@ -46,6 +41,29 @@ func (p *Persister) UpdateClient(ctx context.Context, cl *client.Client) error {
 	})
 }
 
+func (p *Persister) Authenticate(ctx context.Context, id string, secret []byte) (*client.Client, error) {
+	c, err := p.GetConcreteClient(ctx, id)
+	if err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	if err := p.r.ClientHasher().Compare(ctx, c.GetHashedSecret(), secret); err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	return c, nil
+}
+
+func (p *Persister) CreateClient(ctx context.Context, c *client.Client) error {
+	h, err := p.r.ClientHasher().Hash(ctx, []byte(c.Secret))
+	if err != nil {
+		return err
+	}
+
+	c.Secret = string(h)
+	return sqlcon.HandleError(p.Connection(ctx).Create(c, "pk"))
+}
+
 func (p *Persister) DeleteClient(ctx context.Context, id string) error {
 	return sqlcon.HandleError(p.Connection(ctx).Destroy(&client.Client{ID: id}))
 }
@@ -62,22 +80,4 @@ func (p *Persister) GetClients(ctx context.Context, limit, offset int) ([]client
 func (p *Persister) CountClients(ctx context.Context) (int, error) {
 	n, err := p.Connection(ctx).Count(&client.Client{})
 	return n, sqlcon.HandleError(err)
-}
-
-func (p *Persister) GetConcreteClient(ctx context.Context, id string) (*client.Client, error) {
-	var cl client.Client
-	return &cl, sqlcon.HandleError(p.Connection(ctx).Find(&cl, id))
-}
-
-func (p *Persister) Authenticate(ctx context.Context, id string, secret []byte) (*client.Client, error) {
-	c, err := p.GetConcreteClient(ctx, id)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	if err := p.r.ClientHasher().Compare(ctx, c.GetHashedSecret(), secret); err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return c, nil
 }
